# src/main/resources/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Performance Orchestrator API",
    "version": "0.1.0",
    "description": "CKP performance orchestration and integration configuration. Obtain the session cookie and CSRF header from /session for mutations. Execution is explicitly enabled through startup configuration."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
    "/session": {
      "get": {
        "summary": "Session and CSRF token",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Malformed input"
          },
          "403": {
            "description": "Origin/CSRF check or upstream permission denied"
          },
          "404": {
            "description": "Record or source not found"
          },
          "409": {
            "description": "Revision, plan, idempotency or environment conflict"
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": []
      }
    },
    "/environments": {
      "get": {
        "summary": "Environment availability",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Malformed input"
          },
          "403": {
            "description": "Origin/CSRF check or upstream permission denied"
          },
          "404": {
            "description": "Record or source not found"
          },
          "409": {
            "description": "Revision, plan, idempotency or environment conflict"
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": []
      }
    },
    "/services": {
      "get": {
        "summary": "Service catalog",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Malformed input"
          },
          "403": {
            "description": "Origin/CSRF check or upstream permission denied"
          },
          "404": {
            "description": "Record or source not found"
          },
          "409": {
            "description": "Revision, plan, idempotency or environment conflict"
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": []
      }
    },
    "/scenarios": {
      "get": {
        "summary": "Scenario templates",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Malformed input"
          },
          "403": {
            "description": "Origin/CSRF check or upstream permission denied"
          },
          "404": {
            "description": "Record or source not found"
          },
          "409": {
            "description": "Revision, plan, idempotency or environment conflict"
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": []
      }
    },
    "/plans/{id}": {
      "get": {
        "summary": "Read immutable plan",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Malformed input"
          },
          "403": {
            "description": "Origin/CSRF check or upstream permission denied"
          },
          "404": {
            "description": "Record or source not found"
          },
          "409": {
            "description": "Revision, plan, idempotency or environment conflict"
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/runs/{id}": {
      "get": {
        "summary": "Run state",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Malformed input"
          },
          "403": {
            "description": "Origin/CSRF check or upstream permission denied"
          },
          "404": {
            "description": "Record or source not found"
          },
          "409": {
            "description": "Revision, plan, idempotency or environment conflict"
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/runs/{id}/artifacts": {
      "get": {
        "summary": "List completed-run artifact metadata",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Malformed input"
          },
          "403": {
            "description": "Origin/CSRF check or upstream permission denied"
          },
          "404": {
            "description": "Record or source not found"
          },
          "409": {
            "description": "Revision, plan, idempotency or environment conflict"
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/runs": {
      "get": {
        "summary": "Filtered run history",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Malformed input"
          },
          "403": {
            "description": "Origin/CSRF check or upstream permission denied"
          },
          "404": {
            "description": "Record or source not found"
          },
          "409": {
            "description": "Revision, plan, idempotency or environment conflict"
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "environment",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "profileId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "outcome",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "since",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/runs/{id}/events": {
      "get": {
        "summary": "Persisted JSON event cursor or reconnecting SSE",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Malformed input"
          },
          "403": {
            "description": "Origin/CSRF check or upstream permission denied"
          },
          "404": {
            "description": "Record or source not found"
          },
          "409": {
            "description": "Revision, plan, idempotency or environment conflict"
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/runs/{id}/cancel": {
      "post": {
        "summary": "Request idempotent cancellation",
        "responses": {
          "202": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Malformed input"
          },
          "403": {
            "description": "Origin/CSRF check or upstream permission denied"
          },
          "404": {
            "description": "Record or source not found"
          },
          "409": {
            "description": "Revision, plan, idempotency or environment conflict"
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Use the exact csrfHeader name and csrfToken from /session with its session cookie."
          }
        ]
      }
    },
    "/runs/{id}/report": {
      "get": {
        "summary": "Performance report",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Malformed input"
          },
          "403": {
            "description": "Origin/CSRF check or upstream permission denied"
          },
          "404": {
            "description": "Record or source not found"
          },
          "409": {
            "description": "Revision, plan, idempotency or environment conflict"
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/runs/{id}/artifacts/{name}": {
      "get": {
        "summary": "Download summary.json or report.html",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Malformed input"
          },
          "403": {
            "description": "Origin/CSRF check or upstream permission denied"
          },
          "404": {
            "description": "Record or source not found"
          },
          "409": {
            "description": "Revision, plan, idempotency or environment conflict"
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Problem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "field",
          "message"
        ]
      }
    }
  }
}